+++ src/org/ods/orchestration/util/DocumentHistory.groovy
@@ -86,6 +86,7 @@ class DocumentHistory {
         return this
     }
 
+    @NonCPS
     Long getVersion() {
         this.latestVersionId
     }
@@ -194,6 +195,7 @@ class DocumentHistory {
 
     }
 
+    @NonCPS
     protected String getSavedDocumentName(String environment = targetEnvironment) {
         def suffix = (documentName) ? '-' + documentName : ''
         return "documentHistory-${environment}${suffix}"
@@ -276,6 +278,7 @@ class DocumentHistory {
         }
     }
 
+    @NonCPS
     private String createRational(DocumentHistoryEntry currentEntry) {
         if (currentEntry.getEntryId() == 1L) {
             return "Initial document version."
@@ -296,7 +299,8 @@ class DocumentHistory {
         }.findAll { it.id != currentEntry.getEntryId() }
         def concurrentVersions = getConcurrentVersions(oldVersionsSimplified, currentEntry.getPreviousProjectVersion())
 
-        if (currentEntry.getPreviousProjectVersion() && oldVersionsSimplified.size() == concurrentVersions.size()) {
+        if (currentEntry.getPreviousProjectVersion() && oldVersionsSimplified.size() == concurrentVersions.size() &&
+            LeVADocumentUtil.isFullDocument(documentName)) {
             throw new RuntimeException('Inconsistent state found when building DocumentHistory. ' +
                 "Project has as previous project version '${currentEntry.getPreviousProjectVersion()}' " +
                 'but no document history containing that ' +
@@ -408,7 +412,8 @@ class DocumentHistory {
     private List<Map> getIssueChangesForVersion(String version, String issueType, Map issues) {
         // Filter chapter issues for this document only
         if (issueType == JiraDataItem.TYPE_DOCS) {
-            issues = issues.findAll { it.value.documents.contains(this.documentName.split('-').first()) }
+            def docType = LeVADocumentUtil.getTypeFromName(this.documentName)
+            issues = issues.findAll { it.value.documents.contains(docType) }
         }
 
         issues.findAll { it.value.versions?.contains(version) }